+++ packages/orq-rc/src/orq_ai_sdk/models/createchunkop.py
@@ -257,7 +257,7 @@ class CreateChunkResponseBody(BaseModel):
     updated_by_id: str
     r"""The unique identifier of the user who updated the chunk"""
 
-    id: Optional[str] = "chunk_01JQNDTEDAAEK5C45EGRTRX5N7"
+    id: Optional[str] = "chunk_01JQNQ1F7B1335YYCFKP6RCYAF"
     r"""Unique identifier for the element"""
 
     enabled: Optional[bool] = True