+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQNDTED81QAJ4X2164MP45WR"
+        "01JQNQ1F7A1AVCYMKC0B8ZNB25"
     )
     r"""The id of the resource"""